feat: 사이드바 및 메인 앱 구조 개선
- 사이드바 컴포넌트를 별도의 앱으로 분리하여 Shadow DOM에서 렌더링
