+++ src/entrypoints/content.tsx
@@ -1,16 +1,40 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from '../content/App';
+import SidebarApp from '../content/SidebarApp';
+import '../content/styles.css'; // Import CSS for shadow DOM
 
 export default defineContentScript({
   matches: ['<all_urls>'],
-  main() {
+  cssInjectionMode: 'ui', // Enable UI mode for shadow DOM support
+
+  async main(ctx) {
     // WXT Analytics는 자동 초기화됨
 
-    const root = document.createElement('div');
-    root.id = 'tyquill-content-root';
-    document.body.appendChild(root);
+    // Create separate shadow root UI for the main app (FloatingButton)
+    const mainUI = await createShadowRootUi(ctx, {
+      name: 'tyquill-main-ui',
+      position: 'overlay',
+      onMount: (container) => {
+        // Create React root and render the main app inside shadow DOM
+        const reactRoot = createRoot(container);
+        reactRoot.render(<App />);
+      },
+    });
 
-    createRoot(root).render(<App />);
+    // Create separate shadow root UI for the sidebar
+    const sidebarUI = await createShadowRootUi(ctx, {
+      name: 'tyquill-sidebar-ui',
+      position: 'overlay',
+      onMount: (container) => {
+        // Create React root and render the sidebar app inside shadow DOM
+        const reactRoot = createRoot(container);
+        reactRoot.render(<SidebarApp />);
+      },
+    });
+
+    // Mount both UIs
+    mainUI.mount();
+    sidebarUI.mount();
   },
 });
